backend/worker/requirements.txt
numpy>=1.20.0
opencv-python>=4.11.0.86
onnxruntime>=1.8.0
Pillow>=9.0.0
paho-mqtt
psutil
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
twilio>=8.0.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
